+++ Realm/ObjectServerTests/SwiftObjectServerTests.swift
@@ -22,10 +22,10 @@ import RealmSwift
 #if swift(>=3.0)
 class SwiftObjectServerTests: SwiftSyncTestCase {
     /// It should be possible to successfully open a Realm configured for sync.
-    func DISABLED_testBasicSwiftSync() {
+    func testBasicSwiftSync() {
         let url = URL(string: "realm://localhost:9080/~/testBasicSync")!
         do {
-            let user = try synchronouslyLogInUser(for: basicCredential(create: true), server: authURL)
+            let user = try synchronouslyLogInUser(for: basicCredential(register: true), server: authURL)
             let realm = try synchronouslyOpenRealm(url: url, user: user)
             XCTAssert(realm.isEmpty, "Freshly synced Realm was not empty...")
         } catch {
@@ -34,9 +34,9 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
     }
 
     /// If client B adds objects to a Realm, client A should see those new objects.
-    func DISABLED_testSwiftRemoteAddObjects() {
+    func testSwiftAddObjects() {
         do {
-            let user = try synchronouslyLogInUser(for: basicCredential(create: isParent), server: authURL)
+            let user = try synchronouslyLogInUser(for: basicCredential(register: isParent), server: authURL)
             let realm = try synchronouslyOpenRealm(url: realmURL, user: user)
             if isParent {
                 user.waitForDownload(toFinish: realmURL)
@@ -60,9 +60,9 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
     }
 
     /// If client B removes objects from a Realm, client A should see those changes.
-    func DISABLED_testSwiftRemoteDeleteObjects() {
+    func testSwiftDeleteObjects() {
         do {
-            let user = try synchronouslyLogInUser(for: basicCredential(create: isParent), server: authURL)
+            let user = try synchronouslyLogInUser(for: basicCredential(register: isParent), server: authURL)
             let realm = try synchronouslyOpenRealm(url: realmURL, user: user)
             if isParent {
                 try realm.write {
@@ -93,7 +93,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
     func testBasicSwiftSync() {
         let url = NSURL(string: "realm://localhost:9080/~/testBasicSync")!
         do {
-            let user = try synchronouslyLogInUser(for: basicCredential(create: true), server: authURL)
+            let user = try synchronouslyLogInUser(for: basicCredential(register: true), server: authURL)
             let realm = try synchronouslyOpenRealm(url: url, user: user)
             XCTAssert(realm.isEmpty, "Freshly synced Realm was not empty...")
         } catch {
@@ -102,9 +102,9 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
     }
 
     /// If client B adds objects to a Realm, client A should see those new objects.
-    func testSwiftRemoteAddObjects() {
+    func testSwiftAddObjects() {
         do {
-            let user = try synchronouslyLogInUser(for: basicCredential(create: isParent), server: authURL)
+            let user = try synchronouslyLogInUser(for: basicCredential(register: isParent), server: authURL)
             let realm = try synchronouslyOpenRealm(url: realmURL, user: user)
             if isParent {
                 user.waitForDownloadToFinish(realmURL)
@@ -128,9 +128,9 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
     }
 
     /// If client B removes objects from a Realm, client A should see those changes.
-    func testSwiftRemoteDeleteObjects() {
+    func testSwiftDeleteObjects() {
         do {
-            let user = try synchronouslyLogInUser(for: basicCredential(create: isParent), server: authURL)
+            let user = try synchronouslyLogInUser(for: basicCredential(register: isParent), server: authURL)
             let realm = try synchronouslyOpenRealm(url: realmURL, user: user)
             if isParent {
                 try realm.write {