+++ Realm/RLMSyncManager_Private.hpp
@@ -20,19 +20,19 @@
 
 #import "RLMSyncUtil_Private.h"
 
-#import "sync_config.hpp"
-#import "sync_metadata.hpp"
+namespace realm {
+enum class SyncSessionError;
+class SyncMetadataManager;
+}
 
-@class RLMSyncUser;
+@class RLMSyncUser, RLMSyncFileManager;
 
 // All private API methods are threadsafe and synchronized, unless denoted otherwise. Since they are expected to be
 // called very infrequently, this should pose no issues.
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface RLMSyncManager () {
-    std::unique_ptr<realm::SyncMetadataManager> _metadata_manager;
-}
+@interface RLMSyncManager ()
 
 @property (nullable, nonatomic, copy) RLMSyncBasicErrorReportingBlock sessionCompletionNotifier;
 
@@ -43,9 +43,6 @@ NS_ASSUME_NONNULL_BEGIN
  (by using the same configuration) will return `nil`.
  */
 - (nullable RLMSyncSession *)sessionForSyncConfiguration:(RLMSyncConfiguration *)config NS_UNAVAILABLE;
-
-/// Reset the singleton instance, and any saved state. Only for use with Realm Object Store tests.
-+ (void)_resetStateForTesting;
 
 - (void)_fireError:(NSError *)error;
 
@@ -59,6 +56,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (NSArray<RLMSyncUser *> *)_allUsers;
 
+- (RLMSyncFileManager *)fileManager;
+
 /**
  Register a user. If an equivalent user has already been registered, the argument is not added to the store, and the
  existing user is returned. Otherwise, the argument is added to the store and `nil` is returned.
@@ -70,6 +69,9 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)_deregisterLoggedOutUser:(RLMSyncUser *)user;
 
 - (nullable RLMSyncUser *)_userForIdentity:(NSString *)identity;
+
+- (void)_handleBindRequestForSyncConfig:(RLMSyncConfiguration *)syncConfig
+                          localFilePath:(NSString *)filePathString;
 
 NS_ASSUME_NONNULL_END