+++ Realm/RLMSyncUser.mm
@@ -53,8 +53,16 @@ using namespace realm;
 
 #pragma mark - static API
 
-+ (NSArray *)all {
++ (NSArray *)allUsers {
     return [[RLMSyncManager sharedManager] _allUsers];
+}
+
++ (RLMSyncUser *)currentUser {
+    NSArray *allUsers = [[RLMSyncManager sharedManager] _allUsers];
+    if (allUsers.count > 1) {
+        @throw RLMException(@"+currentUser cannot be called if more that one valid, logged-in user exists.");
+    }
+    return allUsers.firstObject;
 }
 
 #pragma mark - API
@@ -72,19 +80,19 @@ using namespace realm;
     return nil;
 }
 
-+ (void)authenticateWithCredential:(RLMSyncCredential *)credential
-                     authServerURL:(NSURL *)authServerURL
-                      onCompletion:(RLMUserCompletionBlock)completion {
-    [self authenticateWithCredential:credential
-                       authServerURL:authServerURL
-                             timeout:30
-                        onCompletion:completion];
++ (void)logInWithCredential:(RLMSyncCredential *)credential
+              authServerURL:(NSURL *)authServerURL
+               onCompletion:(RLMUserCompletionBlock)completion {
+    [self logInWithCredential:credential
+                authServerURL:authServerURL
+                      timeout:30
+                 onCompletion:completion];
 }
 
-+ (void)authenticateWithCredential:(RLMSyncCredential *)credential
-                     authServerURL:(NSURL *)authServerURL
-                           timeout:(NSTimeInterval)timeout
-                      onCompletion:(RLMUserCompletionBlock)completion {
++ (void)logInWithCredential:(RLMSyncCredential *)credential
+              authServerURL:(NSURL *)authServerURL
+                    timeout:(NSTimeInterval)timeout
+               onCompletion:(RLMUserCompletionBlock)completion {
     RLMSyncUser *user = [[RLMSyncUser alloc] initWithAuthServer:authServerURL];
     [RLMSyncUser _performLogInForUser:user
                            credential:credential
@@ -209,13 +217,6 @@ using namespace realm;
                                    kRLMSyncAppIDKey: [RLMSyncManager sharedManager].appID,
                                    } mutableCopy];
     NSMutableDictionary *info = [(credential.userInfo ?: @{}) mutableCopy];
-
-    if (credential.provider == RLMIdentityProviderUsernamePassword) {
-        RLMAuthenticationActions actions = [info[kRLMSyncActionsKey] integerValue];
-        if (actions & RLMAuthenticationActionsCreateAccount) {
-            info[kRLMSyncRegisterKey] = @(YES);
-        }
-    }
 
     if ([info count] > 0) {
         // Munge user info into the JSON request.