+++ RealmSwift/Sync.swift
@@ -34,19 +34,26 @@ public typealias SyncUser = RLMSyncUser
  */
 public typealias SyncManager = RLMSyncManager
 
+extension SyncManager {
+#if swift(>=3.0)
+    /// The sole instance of the singleton.
+    public static var shared: SyncManager {
+        return __shared()
+    }
+#else
+    /// The sole instance of the singleton.
+    @nonobjc public static func sharedManager() -> SyncManager {
+        return __sharedManager()
+    }
+#endif
+}
+
 /**
  A session object which represents communication between the client and server for a specific Realm.
 
  - see: `RLMSyncSession`
  */
 public typealias SyncSession = RLMSyncSession
-
-/**
- An options type which represents certain authentication actions that can be associated with certain credential types.
-
- - see: `RLMAuthenticationActions`
- */
-public typealias AuthenticationActions = RLMAuthenticationActions
 
 /**
  A closure type for a closure which can be set on the `SyncManager` to allow errors to be reported to the application.
@@ -85,13 +92,55 @@ public typealias SyncLogLevel = RLMSyncLogLevel
  */
 public typealias Provider = RLMIdentityProvider
 
+/// A `SyncConfiguration` represents configuration parameters for Realms intended to sync with a Realm Object Server.
+public struct SyncConfiguration {
+    /// The `SyncUser` who owns the Realm that this configuration should open.
+    public let user: SyncUser
+
+    /**
+     The URL of the Realm on the Realm Object Server that this configuration should open.
+
+     - warning: The URL must be absolute (e.g. `realms://example.com/~/foo`), and cannot end with
+                `.realm`, `.realm.lock` or `.realm.management`.
+     */
+    public let realmURL: URL
+
+    /// A policy that determines what should happen when all references to Realms opened by this configuration
+    /// go out of scope.
+    internal let stopPolicy: RLMSyncStopPolicy
+
+    internal init(config: RLMSyncConfiguration) {
+        self.user = config.user
+        self.realmURL = config.realmURL
+        self.stopPolicy = config.stopPolicy
+    }
+
+    func asConfig() -> RLMSyncConfiguration {
+        let config = RLMSyncConfiguration(user: user, realmURL: realmURL)
+        config.stopPolicy = stopPolicy
+        return config
+    }
+
+    /**
+     Initialize a sync configuration with a user and a Realm URL.
+
+     - warning: The URL must be absolute (e.g. `realms://example.com/~/foo`), and cannot end with
+                `.realm`, `.realm.lock` or `.realm.management`.
+     */
+    public init(user: SyncUser, realmURL: URL) {
+        self.user = user
+        self.realmURL = realmURL
+        self.stopPolicy = .afterChangesUploaded
+    }
+}
+
 /// A `Credential` represents data that uniquely identifies a Realm Object Server user.
 public struct Credential {
     public typealias Token = String
 
-    var token: Token
-    var provider: Provider
-    var userInfo: [String: Any]
+    internal var token: Token
+    internal var provider: Provider
+    internal var userInfo: [String: Any]
 
     /// Initialize a new credential using a custom token, authentication provider, and user information dictionary. In
     /// most cases, the convenience initializers should be used instead.
@@ -125,8 +174,8 @@ public struct Credential {
     /// Initialize a new credential using a Realm Object Server username and password.
     public static func usernamePassword(username: String,
                                         password: String,
-                                        actions: AuthenticationActions) -> Credential {
-        return Credential(RLMSyncCredential(username: username, password: password, actions: actions))
+                                        register: Bool = false) -> Credential {
+        return Credential(RLMSyncCredential(username: username, password: password, register: register))
     }
 
     /// Initialize a new credential using a Realm Object Server access token.
@@ -144,14 +193,28 @@ extension RLMSyncCredential {
 extension SyncUser {
     /// Given a credential and server URL, log in a user and asynchronously return a `SyncUser` object which can be used to
     /// open Realms and Sessions.
-    public static func authenticate(with credential: Credential,
-                                    server authServerURL: URL,
-                                    timeout: TimeInterval = 30,
-                                    onCompletion completion: @escaping UserCompletionBlock) {
-        return SyncUser.__authenticate(with: RLMSyncCredential(credential),
-                                   authServerURL: authServerURL,
-                                   timeout: timeout,
-                                   onCompletion: completion)
+    public static func logIn(with credential: Credential,
+                             server authServerURL: URL,
+                             timeout: TimeInterval = 30,
+                             onCompletion completion: @escaping UserCompletionBlock) {
+        return SyncUser.__logIn(with: RLMSyncCredential(credential),
+                                authServerURL: authServerURL,
+                                timeout: timeout,
+                                onCompletion: completion)
+    }
+
+    /// An array of all valid, logged-in users.
+    public static var all: [SyncUser] {
+        return __allUsers()
+    }
+
+    /**
+     The logged-in user. `nil` if none exists.
+
+     - warning: Throws an Objective-C exception if more than one logged-in user exists.
+     */
+    public static var current: SyncUser? {
+        return __current()
     }
 }
 
@@ -164,13 +227,55 @@ extension SyncUser {
  */
 public typealias Provider = String // `RLMIdentityProvider` imports as `NSString`
 
+/// A `SyncConfiguration` represents configuration parameters for Realms intended to sync with a Realm Object Server.
+public struct SyncConfiguration {
+    /// The `SyncUser` who owns the Realm that this configuration should open.
+    public let user: SyncUser
+
+    /**
+     The URL of the Realm on the Realm Object Server that this configuration should open.
+
+     - warning: The URL must be absolute (e.g. `realms://example.com/~/foo`), and cannot end with
+                `.realm`, `.realm.lock` or `.realm.management`.
+     */
+    public let realmURL: NSURL
+
+    /// A policy that determines what should happen when all references to Realms opened by this configuration
+    /// go out of scope.
+    internal let stopPolicy: RLMSyncStopPolicy
+
+    internal init(config: RLMSyncConfiguration) {
+        self.user = config.user
+        self.realmURL = config.realmURL
+        self.stopPolicy = config.stopPolicy
+    }
+
+    func asConfig() -> RLMSyncConfiguration {
+        let config = RLMSyncConfiguration(user: user, realmURL: realmURL)
+        config.stopPolicy = stopPolicy
+        return config
+    }
+
+    /**
+     Initialize a sync configuration with a user and a Realm URL.
+
+     - warning: The URL must be absolute (e.g. `realms://example.com/~/foo`), and cannot end with
+                `.realm`, `.realm.lock` or `.realm.management`.
+     */
+    public init(user: SyncUser, realmURL: NSURL) {
+        self.user = user
+        self.realmURL = realmURL
+        self.stopPolicy = .AfterChangesUploaded
+    }
+}
+
 /// A `Credential` represents data that uniquely identifies a Realm Object Server user.
 public struct Credential {
     public typealias Token = String
 
-    var token: Token
-    var provider: Provider
-    var userInfo: [String: AnyObject]
+    internal var token: Token
+    internal var provider: Provider
+    internal var userInfo: [String: AnyObject]
 
     // swiftlint:disable valid_docs
 
@@ -206,8 +311,8 @@ public struct Credential {
     /// Initialize a new credential using a Realm Object Server username and password.
     public static func usernamePassword(username: String,
                                         password: String,
-                                        actions: AuthenticationActions) -> Credential {
-        return Credential(RLMSyncCredential(username: username, password: password, actions: actions))
+                                        register: Bool = false) -> Credential {
+        return Credential(RLMSyncCredential(username: username, password: password, register: register))
     }
 
     /// Initialize a new credential using a Realm Object Server access token.
@@ -222,18 +327,31 @@ extension RLMSyncCredential {
     }
 }
 
-
 extension SyncUser {
     /// Given a credential and server URL, log in a user and asynchronously return a `SyncUser` object which can be used to
     /// open Realms and Sessions.
-    public static func authenticateWithCredential(credential: Credential,
-                                                  authServerURL: NSURL,
-                                                  timeout: NSTimeInterval = 30,
-                                                  onCompletion completion: UserCompletionBlock) {
-        return SyncUser.__authenticateWithCredential(RLMSyncCredential(credential),
-                                                 authServerURL: authServerURL,
-                                                 timeout: timeout,
-                                                 onCompletion: completion)
+    public static func logInWithCredential(credential: Credential,
+                                           authServerURL: NSURL,
+                                           timeout: NSTimeInterval = 30,
+                                           onCompletion completion: UserCompletionBlock) {
+        return __logInWithCredential(RLMSyncCredential(credential),
+                                     authServerURL: authServerURL,
+                                     timeout: timeout,
+                                     onCompletion: completion)
+    }
+
+    /// An array of all valid, logged-in users.
+    @nonobjc public static func allUsers() -> [SyncUser] {
+        return __allUsers()
+    }
+
+    /**
+     The logged-in user. `nil` if none exists.
+
+     - warning: Throws an Objective-C exception if more than one logged-in user exists.
+     */
+    @nonobjc public static func currentUser() -> SyncUser? {
+        return __currentUser()
     }
 }